requirements.txt
# App
argon2-cffi==21.3.0
asyncpg==0.25.0
alembic==1.7.5
fastapi==0.70.0
pydantic[email]==1.8.2
uvicorn[standard]==0.16.0
sqlalchemy[asyncio,mypy]==1.4.27

# Tooling
asgi-lifespan==1.0.1
black==21.12b0
flake8==4.0.1
httpx==0.21.1
isort==5.10.1
mypy==0.910
psycopg2-binary==2.9.2
pytest==6.2.5
pytest-asyncio==0.16.0
sqlalchemy-utils==0.37.9
